Follow binary references in expunge
